Look-up slack chat window again after timeout
This change addresses #178, which is a ticket concerning constant daily
crashes of pidgin.

The crashes seem to happen when a chat window is closed while a "user is
typing" timeout window is open. When this happens, the chat pointer
stored is freed by pidgin. Then it is dereferenced and crashes.

The change here is to look the chat up again at the end of the timeout.
That way a null returned pointer informs the program that the chat is
no longer open.

The change also gets rid of what I believe to be a double-free.